[IMP] ir.model: new field table_name, allow deletion of table and views that are not in the registry

+++ openerp/addons/base/ir/ir_model.py
@@ -98,6 +98,7 @@ class ir_model(osv.osv):
     _columns = {
         'name': fields.char('Model Description', translate=True, required=True),
         'model': fields.char('Model', required=True, select=1),
+        'table_name': fields.char('Table', required=True, readonly=True),
         'info': fields.text('Information'),
         'field_id': fields.one2many('ir.model.fields', 'model_id', 'Fields', required=True),
         'state': fields.selection([('manual','Custom Object'),('base','Base Object')],'Type',readonly=True),
@@ -145,13 +146,12 @@ class ir_model(osv.osv):
 
     def _drop_table(self, cr, uid, ids, context=None):
         for model in self.browse(cr, uid, ids, context):
-            model_pool = self.pool[model.model]
-            cr.execute('select relkind from pg_class where relname=%s', (model_pool._table,))
+            cr.execute('select relkind from pg_class where relname=%s', (model.table_name,))
             result = cr.fetchone()
             if result and result[0] == 'v':
-                cr.execute('DROP view %s' % (model_pool._table,))
+                cr.execute('DROP view %s' % (model.table_name,))
             elif result and result[0] == 'r':
-                cr.execute('DROP TABLE %s' % (model_pool._table,))
+                cr.execute('DROP TABLE %s' % (model.table_name,))
         return True
 
     def unlink(self, cr, user, ids, context=None):
@@ -185,10 +185,13 @@ class ir_model(osv.osv):
         return super(ir_model,self).write(cr, user, ids, vals, context)
 
     def create(self, cr, user, vals, context=None):
-        if  context is None:
+        if context is None:
             context = {}
         if context and context.get('manual'):
-            vals['state']='manual'
+            vals['state'] = 'manual'
+            table = vals.get('model', '').replace('.', '_')
+            vals['table_name'] = table
+
         res = super(ir_model,self).create(cr, user, vals, context)
         if vals.get('state','base')=='manual':
             self.instanciate(cr, user, vals['model'], context)
@@ -304,14 +307,16 @@ class ir_model_fields(osv.osv):
         for field in self.browse(cr, uid, ids, context):
             if field.name in MAGIC_COLUMNS:
                 continue
-            model = self.pool[field.model]
-            cr.execute('select relkind from pg_class where relname=%s', (model._table,))
+            table_name = field.model_id.table_name
+            cr.execute('select relkind from pg_class where relname=%s', (table_name,))
             result = cr.fetchone()
-            cr.execute("SELECT column_name FROM information_schema.columns WHERE table_name ='%s' and column_name='%s'" %(model._table, field.name))
+            cr.execute("SELECT column_name FROM information_schema.columns WHERE table_name=%s and column_name=%s", (table_name, field.name))
             column_name = cr.fetchone()
             if column_name and (result and result[0] == 'r'):
-                cr.execute('ALTER table "%s" DROP column "%s" cascade' % (model._table, field.name))
-            model._columns.pop(field.name, None)
+                cr.execute('ALTER table "%s" DROP column "%s" cascade' % (table_name, field.name))
+            model = self.pool.get(field.model)
+            if model:
+                model._columns.pop(field.name, None)
         return True
 
     def unlink(self, cr, user, ids, context=None):
@@ -490,7 +495,7 @@ class ir_model_constraint(Model):
         'name': fields.char('Constraint', required=True, select=1,
             help="PostgreSQL constraint or foreign key name."),
         'model': fields.many2one('ir.model', string='Model',
-            required=True, select=1),
+            required=True, select=1, ondelete='cascade'),
         'module': fields.many2one('ir.module.module', string='Module',
             required=True, select=1),
         'type': fields.char('Constraint Type', required=True, size=1, select=1,
@@ -520,7 +525,7 @@ class ir_model_constraint(Model):
         ids.reverse()
         for data in self.browse(cr, uid, ids, context):
             model = data.model.model
-            model_obj = self.pool[model]
+            model_table = data.model.table_name
             name = openerp.tools.ustr(data.name)
             typ = data.type
 
@@ -534,17 +539,17 @@ class ir_model_constraint(Model):
             if typ == 'f':
                 # test if FK exists on this table (it could be on a related m2m table, in which case we ignore it)
                 cr.execute("""SELECT 1 from pg_constraint cs JOIN pg_class cl ON (cs.conrelid = cl.oid)
-                              WHERE cs.contype=%s and cs.conname=%s and cl.relname=%s""", ('f', name, model_obj._table))
+                              WHERE cs.contype=%s and cs.conname=%s and cl.relname=%s""", ('f', name, model_table))
                 if cr.fetchone():
-                    cr.execute('ALTER TABLE "%s" DROP CONSTRAINT "%s"' % (model_obj._table, name),)
+                    cr.execute('ALTER TABLE "%s" DROP CONSTRAINT "%s"' % (model_table, name),)
                     _logger.info('Dropped FK CONSTRAINT %s@%s', name, model)
 
             if typ == 'u':
                 # test if constraint exists
                 cr.execute("""SELECT 1 from pg_constraint cs JOIN pg_class cl ON (cs.conrelid = cl.oid)
-                              WHERE cs.contype=%s and cs.conname=%s and cl.relname=%s""", ('u', name, model_obj._table))
+                              WHERE cs.contype=%s and cs.conname=%s and cl.relname=%s""", ('u', name, model_table))
                 if cr.fetchone():
-                    cr.execute('ALTER TABLE "%s" DROP CONSTRAINT "%s"' % (model_obj._table, name),)
+                    cr.execute('ALTER TABLE "%s" DROP CONSTRAINT "%s"' % (model_table, name),)
                     _logger.info('Dropped CONSTRAINT %s@%s', name, model)
 
         self.unlink(cr, uid, ids, context)
@@ -559,7 +564,7 @@ class ir_model_relation(Model):
         'name': fields.char('Relation Name', required=True, select=1,
             help="PostgreSQL table name implementing a many2many relation."),
         'model': fields.many2one('ir.model', string='Model',
-            required=True, select=1),
+            required=True, select=1, ondelete='cascade'),
         'module': fields.many2one('ir.module.module', string='Module',
             required=True, select=1),
         'date_update': fields.datetime('Update Date'),
@@ -579,7 +584,6 @@ class ir_model_relation(Model):
         ids.sort()
         ids.reverse()
         for data in self.browse(cr, uid, ids, context):
-            model = data.model
             name = openerp.tools.ustr(data.name)
 
             # double-check we are really going to delete all the owners of this schema element
@@ -1181,7 +1185,8 @@ class ir_model_data(osv.osv):
             for (model, res_id, module, name) in to_unlink:
                 if model in self.pool:
                     _logger.info('Deleting %s@%s (%s.%s)', res_id, model, module, name)
-                    self.pool[model].unlink(cr, uid, [res_id], context=context)
+                    if self.pool[model].exists(cr, uid, [res_id], context=context):
+                        self.pool[model].unlink(cr, uid, [res_id], context=context)
 
 class wizard_model_menu(osv.osv_memory):
     _name = 'wizard.ir.model.menu.create'
